third_party/blink/web_tests/svg/dynamic-updates/SVGFECompositeElement-svgdom-k1-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'k1' property of the SVGFECompositeElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Math.round(overComposite1.k1.baseVal * 1000) is 1900
PASS Math.round(overComposite1.k1.baseVal * 1000) is 500
PASS successfullyParsed is true

TEST COMPLETE
